fix(build): Fixed import of a component with a typo. Renamed the component
