+++ src/lib/ui/app/AIChatbot/AskForInsightButton.svelte
