+++ src/lib/ui/app/AIChatbot/index.ts
@@ -1,3 +1,3 @@
 export { default } from './Chatbot.svelte'
-export { default as AiInsightButton } from './AiInsightButton.svelte'
+export { default as AskForInsightButton } from './AskForInsightButton.svelte'
 export { useChatContext } from './ctx.svelte.js'
