+++ src/stories/Design System - App UI/Chart/ChartWidget.svelte
@@ -2,7 +2,7 @@
   import { useTimeZoneCtx } from '$lib/ctx/time/index.js'
   import { useItemViewportPriorityFlow } from '$lib/ctx/viewport-priority/index.js'
   import { getFormattedDetailedTimestamp } from '$lib/utils/dates/index.js'
-  import { AiInsightButton } from '$ui/app/AIChatbot/index.js'
+  import { AskForInsightButton } from '$ui/app/AIChatbot/index.js'
   import { useMetricSeriesCtx } from '$ui/app/Chart/ctx/index.js'
   import BaseChart, {
     ViewportChart,
@@ -62,7 +62,7 @@
 
     <SpikeExplanations>
       {#snippet children({ slug, explanation })}
-        <AiInsightButton {slug} {explanation} />
+        <AskForInsightButton {slug} {explanation}></AskForInsightButton>
       {/snippet}
     </SpikeExplanations>
